+++ packages/powersync/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add `SyncOptions.checkpointMode`, which can be used to replace a legacy implementation for write checkpoints
   with a newer and more efficient endpoint.
+- Add `PowerSyncDatabase.requestCheckpoint`, which can be used to "sync now": The checkpoint completes once all
+  source data created before the checkpoint has synced.
 - Reduce log noise from attachments service ([#408](https://github.com/powersync-ja/powersync.dart/issues/408)).
 - Fix errors that can't be sent across isolates disrupting the sync isolate.
 - Update PowerSync SQLite core extension to 0.5.3, improving error messages for failed internal SQL statements.